Move retro-motion fns to utils
